% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pfi_ccbio}
\alias{pfi_ccbio}
\title{Cover crop biomass production for each year/field}
\format{A data frame with 56 rows and 6 variables:
\describe{
\item{site_name}{Boyd, Stout, Funcke; the name of the site, the link to pfi_eus}
\item{field}{B42, B44, S, F; id of the field within the site, only Boyd had multiple fields}
\item{sys_trt}{silage, grain}
\item{cc_trt}{no, rye}
\item{year}{the year the spring cover crop biomass was measured}
\item{ccbio_Mgha}{the amount of cover crop biomass in Mg/ha; 1000 kg = 1 Mg}

...
}}
\usage{
pfi_ccbio
}
\description{
Cover crop biomass production for each year/field
}
\keyword{datasets}
