% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pfi_eus}
\alias{pfi_eus}
\title{Unique identifiers of the experimental units and blocks}
\format{A data frame with 56 rows and 7 variables:
\describe{
\item{site_name}{Boyd, Stout, Funcke; name of the physical site}
\item{field}{Boyd42, Boyd44, Stout, Funcke; name of the field within the site}
\item{sys_trt}{silage, grain}
\item{cc_trt}{no, rye}
\item{crop_2019}{corn, soy; the crop that was planted in spring of 2019}
\item{rep}{Numeric rep, defined within a site}
\item{repID}{A concatonation of the field identifier, the sys_trt, underscore numeric rep. This creates a unique identifier for each block in each field at each site.}
...
}}
\usage{
pfi_eus
}
\description{
A dataset containing information about each experimental unit of the research
}
\keyword{datasets}
