% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pfi_ghobsraw}
\alias{pfi_ghobsraw}
\title{Raw weed counts}
\format{A data frame with 430 rows and 31 variables:
\describe{
\item{obs_date}{date, in yyy-mm-dd format, the weed counts were recorded}
\item{obs_initials}{initials of person who observed the data}
\item{electrec_initials}{initals of person who electronically scribed the data}
\item{site_name}{Boyd, Stout, Funcke; the name of the site, the link to pfi_eus}
\item{field}{B42, B44, S, F; name of the field within the site}
\item{sys_trt}{silage, grain}
\item{cc_trt}{no, rye}
\item{crop_2019}{corn, soy; crop grown in 2019 on that field}
\item{rep}{the field replicate}
\item{blockID}{the unique identifier for the block}
\item{trayID}{the unique identifier for the tray; in general each eu had 3 trays}
\item{AMATU}{the weed code for the observed weed, see pfi_weedsplist}

...
}}
\usage{
pfi_ghobsraw
}
\description{
A dataset containing all of the observations made on the trays, weed names have NOT been fixed
}
\keyword{datasets}
