% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pfi_mccbio}
\alias{pfi_mccbio}
\title{Mean cover crop biomass production for each field}
\format{A data frame with 6 rows and 5 variables:
\describe{
\item{site_name}{Boyd, Stout, Funcke; the name of the site, the link to pfi_eus}
\item{field}{B42, B44, S, F; name of the field within the site}
\item{sys_trt}{silage, grain}
\item{cc_trt}{no, rye}
\item{mccbio_Mgha}{the amount of cover crop biomass averaged over all years in Mg/ha; 1000 kg = 1 Mg}

...
}}
\usage{
pfi_mccbio
}
\description{
Mean cover crop biomass production for each field
}
\keyword{datasets}
