% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pfi_siteinfo}
\alias{pfi_siteinfo}
\title{Information about the cooperators and general site info}
\format{A data frame with 4 rows and 6 variables:
\describe{
\item{coop_name}{Kohler, Stout, Funcke; the last name of the cooperator, matches with pfi_siteinfo data.}
\item{sys_trt}{silage, grain}
\item{city}{Closest Iowan city to site}
\item{lat}{Latidude in decimal form, N}
\item{lon}{Longitude in decimal form, W}
\item{county}{County of site}

...
}}
\usage{
pfi_siteinfo
}
\description{
A dataset containing information about the cooperators and general site info
}
\keyword{datasets}
