% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pfi_weedsplist}
\alias{pfi_weedsplist}
\title{Weed abbreviations and descriptions}
\format{A data frame with 18 rows and 7 variables:
\describe{
\item{common_name}{common name of weed}
\item{scientific_name}{scientific name of weed, genus and species}
\item{code}{weed code}
\item{possible_mislabel}{corrections made to original dataset, corrected version appears in pfi_ghobsraw data}
\item{photo_path}{photosynthetic pathway, C3 or C4}
\item{functional_grp}{functional group of weed, forb or grass}
\item{family}{weed family}

...
}}
\usage{
pfi_weedsplist
}
\description{
Weed abbreviations and descriptions
}
\keyword{datasets}
