% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_tom2.R
\name{pfifun_seedstom2conv}
\alias{pfifun_seedstom2conv}
\title{pfifun_seedstom2conv}
\usage{
pfifun_seedstom2conv(totseeds)
}
\arguments{
\item{totseeds}{Number of seeds found in a plot}
}
\value{
Seeds per plot converted to seeds per m2 assuming 20 cores per plot
}
\description{
pfifun_seedstom2conv
}
