% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_ghobs.R
\name{pfifun_sum_byeu}
\alias{pfifun_sum_byeu}
\title{pfifun_sum_byeu}
\usage{
pfifun_sum_byeu(rawdata)
}
\arguments{
\item{rawdata}{Tibble of the form pfi_ghobsraw}
}
\value{
A tibble with total seeds per eu
}
\description{
pfifun_sum_byeu
}
