% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_ghobs.R
\name{pfifun_sum_eu}
\alias{pfifun_sum_eu}
\title{pfifun_sum_eu}
\usage{
pfifun_sum_eu(rawdata)
}
\arguments{
\item{rawdata}{Consider using pfi_ghobsraw}
}
\value{
A tibble with total seeds per eu
}
\description{
pfifun_sum_eu
}
