% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_ghobs.R
\name{pfifun_sum_weed}
\alias{pfifun_sum_weed}
\title{pfifun_sum_weed}
\usage{
pfifun_sum_weed(rawdata)
}
\arguments{
\item{rawdata}{Defaults to pfi_ghobsraw}
}
\value{
A tibble with total seeds by weed and experimental unit (eu)
}
\description{
pfifun_sum_weed
}
