% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_ghobs.R
\name{pfifun_sum_weedbyeu}
\alias{pfifun_sum_weedbyeu}
\title{pfifun_sum_weedbyeu}
\usage{
pfifun_sum_weedbyeu(rawdata)
}
\arguments{
\item{rawdata}{Tibble of the form pfi_ghobsraw}
}
\value{
A tibble with total seeds by weed and experimental unit (eu)
}
\description{
pfifun_sum_weedbyeu
}
