% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_ghobs.R
\name{pfifun_sumeu}
\alias{pfifun_sumeu}
\title{pfifun_sumeu}
\usage{
pfifun_sumeu(sumdata)
}
\arguments{
\item{sumdata}{Output from pfi_sumwdsp function}
}
\value{
A tibble with total seeds per eu
}
\description{
pfifun_sumeu
}
