% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_ghobs.R
\name{pfifun_sumwdsp}
\alias{pfifun_sumwdsp}
\title{pfifun_sumwdsp}
\usage{
pfifun_sumwdsp(rawdata)
}
\arguments{
\item{rawdata}{Defaults to pfi_ghobsraw}
}
\value{
A tibble with total seeds by weed and experimental unit (eu)
}
\description{
pfifun_sumwdsp
}
