% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pfi_ghobsraw}
\alias{pfi_ghobsraw}
\title{Raw weed counts}
\format{A data frame with 476 rows and 27 variables:
\describe{
\item{obs_date}{date, in yyy-mm-dd format, the weed counts were recorded}
\item{site_name}{Boyd, Stout, Funcke; the name of the site, the link to pfi_eus}
\item{field}{B42, B44, S, F; name of the field within the site}
\item{sys_trt}{silage, grain}
\item{cc_trt}{no, rye}
\item{rep}{the field replicate}
\item{repID}{the unique identifier for the replicate}
\item{AMATU}{the weed code for the observed weed, see pfi_weedsplist}

...
}}
\usage{
pfi_ghobsraw
}
\description{
A dataset containing all of the observations made on the trays
}
\keyword{datasets}
